move announcements to course
